requirements.txt
python-dateutil
aiohttp
discord.py
quart
